Fix bool props (#132)
Add missing boolean properties and fix SSR boolean prop test.

The test for boolean properties only checks that the property name _includes_ a boolean property key, not that the actual property is a match, start to end. This means that attributes such as `aria-selected=${true}` and `data-selected=${true}` would be matched and either removed or have their value replaced with `selected`.