Use acorn-node
Uses acorn 7, with support for new ES features that have already landed
in Node.js like bigint and numeric separators.

The breaking change in acorn does not affect end users of nanohtml. It's
just a node name change: Import is now ImportExpression.

acorn-node@1.x patches acorn to undo the breaking change, so users of that
version can upgrade to acorn 7 with no problems. In 2.x that patch will
be removed, but the breaking change would still not affect nanohtml,
because we don't deal with dynamic imports anyway :)